kyc-worker/requirements.txt
pika==1.3.2
requests==2.32.3
python-dotenv==1.0.1
uvloop==0.20.0; platform_system != 'Windows'
opencv-python-headless==4.10.0.84
numpy==1.26.4
onnxruntime==1.20.0
Pillow==10.4.0

